# docs/swagger.yaml
definitions:
...
  dto.ProductColorDTO:
    properties:
      color_name:
        type: string
      color_value:
        type: string
    type: object
  dto.ProductCreateDTO:
    properties:
      Product_name:
        type: string
      brand_id:
        type: integer
      category_id:
        type: integer
      description:
        type: string
      is_new:
        type: integer
      price:
        type: integer
      product_color:
        items:
          $ref: '#/definitions/dto.ProductColorDTO'
        type: array
      product_image:
        items:
          $ref: '#/definitions/dto.ProductImageDTO'
        type: array
      product_size:
        items:
          $ref: '#/definitions/dto.ProductSizeDTO'
        type: array
      rating:
        type: integer
      stock:
        type: integer
      store_id:
        type: integer
    required:
    - Product_name
    - brand_id
    - category_id
    - description
    - is_new
    - price
    - stock
    - store_id
    type: object
  dto.ProductImageDTO:
    properties:
      photo:
        type: string
    type: object
  dto.ProductSizeDTO:
    properties:
      size:
        type: string
    type: object
  dto.ProductUpdateDTO:
    properties:
      Product_name:
        type: string
      brand_id:
        type: integer
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      is_new:
        type: integer
      price:
        type: integer
      product_color:
        items:
          $ref: '#/definitions/dto.ProductColorDTO'
        type: array
      product_image:
        items:
          $ref: '#/definitions/dto.ProductImageDTO'
        type: array
      product_size:
        items:
          $ref: '#/definitions/dto.ProductSizeDTO'
        type: array
      rating:
        type: integer
      stock:
        type: integer
      store_id:
        type: integer
    required:
    - Product_name
    - brand_id
    - category_id
    - description
    - is_new
    - price
    - stock
    - store_id
    type: object
...
  helper.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /product:
    get:
      consumes:
      - application/json
      description: Get list products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
      summary: List products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Post product
      parameters:
      - description: the body to create product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/dto.ProductCreateDTO'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Store product
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Delete product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get detail product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Detail product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/dto.ProductUpdateDTO'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - BearerToken: []
      summary: Update product
      tags:
      - Product
...
swagger: "2.0"